pyproject.toml
[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = "true"

[project]
name = "livaria"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "fujulia", email = "juliafu713@gmail.com"},
]
dependencies = [
    "django>=5.0.4",
    "djangorestframework>=3.15.1",
    "django-cors-headers>=4.3.1",
    "djangorestframework-simplejwt>=5.3.1",
    "setuptools>=69.2.0",
    "python-magic>=0.4.27",
    "Pillow>=10.2.0",
    "drf-spectacular>=0.27.1",
    "python-dotenv>=1.0.1",
    "netifaces>=0.11.0",
    "gunicorn>=21.2.0",
    "whitenoise[brotli]>=6.6.0",
    "psycopg2-binary>=2.9.9",
    "django-cloudinary-storage>=0.3.0",
    "django-filter>=24.2",
]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm]
distribution = false

[tool.pdm.scripts]
pre_dev = "python ./scripts/set_my_ip.py"
dev = "python manage.py runserver 0.0.0.0:19003"


